+++ src/dark/colors/editorColors.js
@@ -52,5 +52,5 @@ module.exports = {
   "editorGutter.background": theme.background,
   "editorGutter.modifiedBackground": theme.blue,
   "editorGutter.addedBackground": theme.green,
-  "editorGutter.deletedBackground": theme.red
+  "editorGutter.deletedBackground": theme.red,
 };