+++ src/dark/theme.js
@@ -1,24 +1,31 @@
 module.exports = {
   transparent: "#FFFFFF00",
 
-  background: "#1a1a1a",
-  backgroundDark: "#171C1C",
-  backgroundLight: "#333D3D",
+  background: "#0d1017",
+  backgroundDark: "#0d1117",
+  backgroundLight: "#333D3D", //Border
 
-  foreground: "#D4D4D4",
+  foreground: "#e0Dfd0",
   foregroundDark: "#808080",
 
-  black: "#fff",
-  red: "#E15A60",
-  purple: "#C792EA",
-  orange: "#e1975a",
-  green: "#5abd71",
-  yellow: "#fadc63",
-  blue: "#599ce0",
-  magenta: "#ef6d90",
-  cyan: "#5bd7b9",
+  black: "#000",
   white: "#e0dfdf",
+  gray: "#abb2bf",
+  red: "#E15A60",
+  purple: "#d38efc",
+  orange: "#ffac60",
+  green: "#7cffb4",
+  yellow: "#faf55b",
+  blue: "#3691ff",
+  magenta: "#ff71bc",
+  cyan: "#10b1fe",
 
   // Comentario
   comment: "#414141",
+
+  // punctuation
+  punctuation: "#fff",
+
+  // Colors
+  primary: "#ff0055",
 };