+++ src/light/tokenColors/semanticTokenColors.js
@@ -1,0 +1,9 @@
+const theme = require("../theme");
+
+module.exports = [
+  {
+    operator: "#7a82da",
+    memberOperatorOverload: "#7a82da",
+    operatorOverload: "#7a82da",
+  },
+];