+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## v0.0.4
+
+- Error checking
+- resolve bug
+- change of colors
+- background color change
+- plus
+
 ## v0.0.3
 
 - Error checking